% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{bi_auth}
\alias{bi_auth}
\title{Simple Authentication}
\usage{
bi_auth(email, password, endpoint = "login")
}
\arguments{
\item{email}{A string containing your email used to login.}

\item{password}{A string containing your password used to login.}

\item{endpoint}{A string containing the appropriate endpoint.}
}
\value{
A tibble containing cookie information.
}
\description{
Simple login authentication to the YouGov BrandIndex API.
}
