% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_data.R
\name{company_brands}
\alias{company_brands}
\title{Company Brands}
\usage{
company_brands(region, sector = NULL, custom_sector = NULL, inactive = 0,
  endpoint = "company/brands")
}
\arguments{
\item{region}{A string containing the region for which the brand should be listed.}

\item{sector}{A number containing the sector for which the brands should be listed.}

\item{custom_sector}{A number for the custom sector for which brands should be listed.}

\item{inactive}{A number value to return inactive brands as well (default: 0, 1)}

\item{endpoint}{A string containing the appropriate endpoint.}
}
\value{
A list of brands accessibly by your company, by region and sector.
}
\description{
This endpoint contains the brands accessibel by your company, by region
and sector.
}
\details{
You may list brands by either \emph{sector} or \emph{custom_sector},
but not both.
}
