% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_data.R
\name{company_custom_sectors}
\alias{company_custom_sectors}
\title{Company Custom Sectors}
\usage{
company_custom_sectors(region, endpoint = "company/custom-sectors")
}
\arguments{
\item{region}{A string containing the region for which custom sectors
should be listed.}

\item{endpoint}{A string containing the appropriate endpoint.}
}
\value{
A list of regions for which the custom sectors should be listed.
}
\description{
Company Custom Sectors
}
