% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_data.R
\name{company_regions}
\alias{company_regions}
\title{Company Regions}
\usage{
company_regions(endpoint = "company/region")
}
\arguments{
\item{endpoint}{A string containing the appropriate endpoint.}
}
\value{
A list of regions accessible by your company.
}
\description{
This endpoint contains the regions accessible by your company.
}
