% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_data.R
\name{company_sectors}
\alias{company_sectors}
\title{Company Sectors}
\usage{
company_sectors(region, custom = 0, inactive = 0,
  endpoint = "company/sectors?region=")
}
\arguments{
\item{region}{A string for which the sectors should be listed.}

\item{custom}{A number value to return expired sectors (default: 0, 1)}

\item{inactive}{A number value to return custom sectors (default: 0, 1)}

\item{endpoint}{A string containing the appropriate endpoint.}
}
\value{
A list of sectors by region.
}
\description{
This endpoint contains the sectors by your company, by region.
}
