% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{composite_filters}
\alias{composite_filters}
\title{Composite Filter}
\usage{
composite_filters(id, endpoint = "filters/composite-filter")
}
\arguments{
\item{id}{A string containing the ID for which the composite filter should be described.}

\item{endpoint}{A string containing the appropriate endpoint.}
}
\value{
An ID for the composite filter described.
}
\description{
This endpoint describes a composite filter that your company is able to use.
}
