% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{demo_filter}
\alias{demo_filter}
\title{Demographic Filter}
\usage{
demo_filter(region, name, endpoint = "filters/demo")
}
\arguments{
\item{region}{A string containing the region for which the demographic filter should be described.}

\item{name}{A string containing the \emph{name} for the demographic filter to be described.}

\item{endpoint}{A string containing the appropriate endpoint.}
}
\value{
A list of filter descriptions for a particular demographic filter.
}
\description{
This endpoint contains the demographic filter description for a particular demographic filter required.
}
