% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{demo_filters}
\alias{demo_filters}
\title{Demographic Filters}
\usage{
demo_filters(region, sector = NULL, custom_sector = NULL,
  endpoint = "filters/demos")
}
\arguments{
\item{region}{A string for which the demographic filters should be listed.}

\item{sector}{A number of the sector for which the demographic filtes should be listed.}

\item{custom_sector}{A number of the custom sector for which the demographic filters should be listed.}

\item{endpoint}{}
}
\value{
A list of demographic filters.
}
\description{
This endpoint contains the demographic filters that your company
is able to use.
}
\details{
The \emph{sector} and \emph{custom_sector} are mutually exclusive.
}
