% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_and_clean}
\alias{get_and_clean}
\title{Cleaning HTTP GET Requests}
\usage{
get_and_clean(url)
}
\arguments{
\item{url}{A string of the url to send the GET request to.}
}
\value{
A response from the GET request (list, tibble, error)
}
\description{
This function sends a GET request to the appropriate URL.
Next, it checks the response and grabs the GET request's
content.
}
