% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{metric_filters}
\alias{metric_filters}
\title{Metric Filters}
\usage{
metric_filters(endpoint = "filters/metrics")
}
\arguments{
\item{endpoint}{A string containing the appropriate endpoint.}
}
\value{
A list of metric filters.
}
\description{
This endpoint contains the metric filters that your company is able to use.
}
