% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\title{Metrics Data}
\usage{
metrics(endpoint = "metrics")
}
\arguments{
\item{endpoint}{A string containing the appropriate endpoint.}
}
\value{
A list of the metrics to use with timelines.
}
\description{
This endpoint lists the available metrics for usage with timelines.
}
