% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{multi_sector_timeline}
\alias{multi_sector_timeline}
\title{Multi Sector Timeline}
\usage{
multi_sector_timeline(sector, custom_sector = NULL, start_date, end_date,
  scoring, moving_average = 1, period_type = "day", period_size = 1,
  endpoint = "timeline/multi-brand-file.csv")
}
\arguments{
\item{sector}{A string of the format \code{region:sector_id[:demo_filters][:metric_filters][:metrics]}.
Can be provided multiple times - if so, multiple sectors will be listed.}

\item{custom_sector}{A string of the format \code{region:custom_sector_id[:demo_filters][:metric_filters][:metrics]}. Mutliple can be listed.}

\item{start_date}{A string in the format \code{YYYY-MM-DD}.}

\item{end_date}{A string in the format \code{YYYY-MM-DD}.}

\item{scoring}{A string containing the scoring option to use (default: "total", "aware", "opinion").}

\item{moving_average}{A number of the moving_average, in days, to be applied.}

\item{period_type}{A string for the type of period, if rollling up the
datapoints (default: "day", "week", "month", "year").}

\item{period_size}{A number for the size of the period (default: 1). \code{2}
for rolling up "every 2 weeks".}

\item{endpoint}{A string containing the appropriate endpoint.}
}
\value{
A tibble for multiple sector timeline datapoins.
}
\description{
This endpoint is similar to simple timeline files, but instead of representing
the data for one brand only it can contaim more than one brand per file.
}
\details{
You must provide one of \code{sector} or \code{custom_sector}, but not both.

It's also possible to roll up the data in periods - for this, see
\code{period_type} and \code{period_size} parameters. In this case, when
periodizing for more than "every 1 day" (which is the default, i.e., no
roll-up done), the first date of each period is used in the resulting data.
}
