% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{sectors_with_brands}
\alias{sectors_with_brands}
\title{Sectors with Brands}
\usage{
sectors_with_brands(region, endpoint = "taxonomy/sectors-with-brands.csv?")
}
\arguments{
\item{region}{A string of the region for which the sectors should be listed.}

\item{endpoint}{A string containing the appropriate endpoint.}
}
\value{
A tibble of sectors with their brands.
}
\description{
This endpoint contains data that exposes the taxonomy of categories,
brands, etc...
}
