% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{single_brand_timeline}
\alias{single_brand_timeline}
\title{Single-Brand Timeline File}
\usage{
single_brand_timeline(region, sector, brand_id, start_date, end_date, scoring,
  moving_average, filters = NULL, metrics = NULL,
  endpoint = "timeline/file.csv")
}
\arguments{
\item{region}{A string for the region to which the brand belongs.}

\item{sector}{A number for the sector to which the brand belongs.}

\item{brand_id}{A number for the brand_id to which the brand belongs.}

\item{start_date}{A string in the format \code{YYYY-MM-DD}.}

\item{end_date}{A string in the format \code{YYYY-MM-DD}.}

\item{scoring}{A string containing the scoring option to use (default: "total", "aware", "opinion").}

\item{moving_average}{A number of the moving_average, in days, to be applied.}

\item{filters}{A string in the format \code{[:demo_filters][:metric_filters]}.}

\item{metrics}{A string containing the specific metrics to include if not all
of them are wanted. If you want more than one separate them with a \code{/}.}

\item{endpoint}{A string containing the appropriate endpoint.}
}
\value{
A tibble of the single-brand datapoint.
}
\description{
This endpoint is actually a file payload with datapoint by data and metric,
so its content can be save to a file on disk or worked on directly.
}
