% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{timeline_brands}
\alias{timeline_brands}
\title{Formulas}
\usage{
timeline_brands(brand, start_date, end_date, scoring, moving_average = 1,
  endpoint = "timeline/brands")
}
\arguments{
\item{brand}{A string in the format \code{region:sector_id:brand_id[:demo_filters]}.}

\item{start_date}{A string in the format \code{YYYY-MM-DD}.}

\item{end_date}{A string in the format \code{YYYY-MM-DD}.}

\item{scoring}{A string containing the scoring option to use (default: "total", "aware", "opinion").}

\item{moving_average}{A number of the moving_average, in days, to be applied.}

\item{endpoint}{A string containing the appropriate endpoint.}
}
\value{
A tibble with timeline data.
}
\description{
Scoring Param    Description           3-Point Measure (Pos/Neg/Neu)       2-Point (yes/no)
total            all respondents       unaware == neu                      unaware == no
aware            no unaware            no unaware                          no unaware
opinion          no unaware or neu     no unaware + neu                    no unaware + neu
Timeline Brands
}
\details{
This endpoint contains the URLs from where timeline series files can be retrieved.
Where indicated, these URLs accept optional parts with demographic and metric
filters (referenced as \code{:demo_filters} and \code{:metric_filters}
here in the documentation) for the reponses used to calculate the datapoints.
\code{:demo_filters} has the following format:
\code{:demo_filter_nickname.filter_value[/another_demo.another_value]} (starting
with a colon), each slash representing the start of a demo filter. \code{:metric_filters}
has the following format:
\code{:/brand.brand_id.metric_nick.filter_value[/brand.another_brand_id.another_metric.nick.another_value]}
(starting with a colon), each slash representing the start of a demo filter. Notice
that each metric filter has to start exactly with the characters \code{brand}
(this pattern is used to differentiate between the two types of filters). You
can also provide multiple values for each filter; to do so, just put the numeric
fitler values as the \code{filter_value} part separated by a comma (",").
For example: \code{:/brand.1234.buzz.1,2/brand.2345.quality.9}.
}
