% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{url_builder}
\alias{url_builder}
\title{Building Up The URL}
\usage{
url_builder(endpoint)
}
\arguments{
\item{endpoint}{A string of the endpoint to query.}
}
\value{
A url with the base url and endpoint appended.
}
\description{
This function holds the base url and takes one endpoint.
}
